# swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SKKU Map API",
    "description": "성균관대학교 캠퍼스맵 API 문서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/search/all/{inputquery}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "inputquery",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/search/detail/{buildNo}/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "buildNo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/search/option3/{inputquery}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "inputquery",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}